requirements.txt
tensorflow
numpy
pandas
opencv-python
fastapi
uvicorn[standard]
python-multipart
matplotlib
pillow
torch
torchvision
timm
scikit-learn
